Update to internal offset call
Internal Offset call is no longer used.  It has been commented out of the InitializeRTD function.  This is due to complications around it's implementation.  To be used properly, the external voltage reference must be used.  To use the external reference, the IDACs must be active and there must be an open flowing path.  For there to be a flowing path, the IC must be switched on (which could be done through coding) and an RTD must be connected at startup to ensure current path.  After some use and testing it seems much simpler to leave this call out as it has no substantial impact to the reading accuracy.